refactor the layout steps
