+++ libs/newsletter-workflow/src/lib/steps/cancelLayout.ts
@@ -1,0 +1,12 @@
+import type { WizardStepLayout } from '@newsletters-nx/state-machine';
+
+const markdownToDisplay = `
+# Cancelled
+
+Creation of the newsletter was cancelled.
+`.trim();
+
+export const cancelLayout: WizardStepLayout = {
+	markdownToDisplay,
+	buttons: {},
+};
