+++ libs/newsletter-workflow/src/lib/steps/finishLayout.ts
@@ -1,0 +1,12 @@
+import type { WizardStepLayout } from '@newsletters-nx/state-machine';
+
+const markdownToDisplay = `# Finished
+
+You have reached the end of the wizard.`;
+
+const finishLayout: WizardStepLayout = {
+	markdownToDisplay,
+	buttons: {},
+};
+
+export { finishLayout };
